requirements.txt
hydra-core==1.3.2
icecream==2.1.3
matplotlib==3.8.0
numpy==1.26.0
omegaconf==2.3.0
Pillow==10.0.1
lightning==2.0.9
rich==13.6.0
rtpt==0.0.4
scikit_learn==1.3.1
timm==0.9.7
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
wandb==0.15.12
hydra_colorlog==1.2.0
einops==0.7.0
SciencePlots==2.1.1
fast_pytorch_kmeans==0.2.0.1
seaborn==0.13.2
tueplots==0.0.13
runitall @ git+https://github.com/braun-steven/runitall.git@61efa04cd44f13ac79285673016d8c14afa2c02f
